Refactor: Service 단 인터페이스 분리
Service 단 인터페이스 생성 후 분리 시켰습니다

Related to: prgrms-be-devcourse/NBB1_2_3_Team10#10
